Pane width math: move to a router-free module, cap the fraction at 1
Two fixes to the fraction model, both in the same arithmetic.

The helpers lived in pane.ts, which imports @platform/lib/router — and the
router reads `location` at MODULE INIT for its embed-prefix constant, so
`bun test .../pane.test.ts` alone threw ReferenceError; the suite only passed
because an earlier file installed a DOM shim. The math moves to
listing/pane-math.ts, which imports nothing. That is the cleaner of the two
fixes: deferring the router's read would make a genuinely load-time constant
lazy for every caller in the app to suit a test, and this arithmetic never
wanted the router in the first place.

dragPaneFrac could also return more than 1: a container narrower than
PANE_MIN_W (panel-split grids, zoom) makes the clamp's floor wider than the
container, so the pane rendered at flexBasis 110% and the saved panew was
thrown away as a "legacy pixel width" on the next open. The fraction is now
capped at 1 — the pane wants the whole container, and .listing-main's CSS
min-width is what still holds the list on screen.
