+++ SPEC.md
@@ -85,7 +85,7 @@ The listing may show the selected entry's preview beside the list — Finder's l
   - a **file** → its default template mode (PT-8/PT-9) in an **iframe**, `/render?path=<template>&_file=<file>`, exactly as the preview view builds it (PT-2) — one code path, so a file previews identically in the pane and full-screen. No mode switcher in the pane; the pane shows the default, and full-screen is a double-click on the row (FS-15). "Default" means PT-9's rule in full, including its tail: the first **unconditional** entry renders immediately (CT-12 — no waiting on a gate), and an **all-conditional** list resolves its gates and shows the first **allowed** one, so the pane never claims "no preview" for a file that opens fine full-screen. A file with no usable entry at all (empty list, or every gate denied/broken — fail closed) gets the metadata card below. **The pane never edits the mode list on account of its own width** (PT-15): it is a narrow host — 220 px at the floor, half the container by default (FS-12) — and a template whose layout needs more than that collapses itself at its own breakpoint, which is why nothing here knows how wide any template wants to be.
   - a **directory** → a **folder peek**, which is *not* a recursive listing-with-a-pane: a folder holding exactly one app entry (D124's `isAppEntry`) embeds that app, and any other folder shows a **read-only mini child list** (names + icons, no sort, no file ops, no search, not navigable). Read-only is deliberate — two live listings on screen means two watch sockets, two selection models and an ambiguous target for a delete, and the pane exists to answer "what is in here" before you commit to going in.
   - **nothing selected** → the **self target**: the pane's subject is the folder *already open on the left*, and it has **no preview at all**. The header carries the folder's icon, its name and its own primary action (the host's "Open as app", passed down — a folder with no single app leaves the slot empty), and the body is the neutral hint `Select a file to preview.` (the `.pane-hint` the empty pane already had). **There is NO mode picker**, and therefore no mode: the self target never builds a mode list, resolves no default, and issues no stat or lone-app probe. The `/` key's peers are heavyweight opt-ins (D235/D237 put the chat on every folder), so a picker here offered a chat on the folder from a header that otherwise said "select something" — a `Choose view` chip pointing at a view nobody came for. The folder's own modes remain one click away on the **left** half (the browse chip / the folder's own view), and every file in the folder previews on selection; since opening a folder now auto-selects its first file (FS-16), this state is mostly an empty folder or one holding only folders. *There used to be an elaborate rule here instead — drop `_listing`, offer the peers, but land on no mode unless the folder had a lone app of its own, with a pane-only `_none` entry before that. Hiding the picker deletes the question those were answering.*
-- **FS-12** **Draggable divider** between list and pane. **The split is a FRACTION of the split container, not a pixel width** — state is `{ on, frac }`, rendered as a percentage `flex-basis`, defaulting to `PANE_DEFAULT_FRAC = 0.5` (half and half). A proportion is what the user actually chose ("about a third of the window to the list"), and it is the only form of the answer that stays right when the window resizes: the previous model resolved a measured half into pixels at mount and then never rescaled, so a pane sized on a wide monitor became the whole view on a laptop. A percentage also needs **no measurement** — it is correct before the first paint, whatever the container turns out to be — so the measuring layout effect and its unmeasurable-container fallback (`PANE_FALLBACK_W = 420`) are both gone. **The pixel floors remain, and they are where the pixels live now:** the drag runs the cursor's distance from the container's right edge through `clampPaneWidth` (pane ≥ 220 px, list ≥ 60 px, the pane's floor applied last so a container too small for both keeps the pane and scrolls the list) and stores `clamped ÷ container width`; `.listing-pane-slot` and `.listing-main` carry the same two floors as CSS `min-width`, which is what holds them when the *window* — not the divider — is what changed. Only a **dragged** fraction persists (FS-13); the default is never written to `panew`.
+- **FS-12** **Draggable divider** between list and pane. **The split is a FRACTION of the split container, not a pixel width** — state is `{ on, frac }`, rendered as a percentage `flex-basis`, defaulting to `PANE_DEFAULT_FRAC = 0.5` (half and half). A proportion is what the user actually chose ("about a third of the window to the list"), and it is the only form of the answer that stays right when the window resizes: the previous model resolved a measured half into pixels at mount and then never rescaled, so a pane sized on a wide monitor became the whole view on a laptop. A percentage also needs **no measurement** — it is correct before the first paint, whatever the container turns out to be — so the measuring layout effect and its unmeasurable-container fallback (`PANE_FALLBACK_W = 420`) are both gone. **The pixel floors remain, and they are where the pixels live now:** the drag runs the cursor's distance from the container's right edge through `clampPaneWidth` (pane ≥ 220 px, list ≥ 60 px, the pane's floor applied last so a container too small for both keeps the pane and scrolls the list) and stores `clamped ÷ container width`, capped at **1** — a container narrower than the pane's own floor would otherwise yield a fraction above 1, which renders as an over-100 % basis and reads back as a legacy pixel value; `.listing-pane-slot` and `.listing-main` carry the same two floors as CSS `min-width`, which is what holds them when the *window* — not the divider — is what changed. Only a **dragged** fraction persists (FS-13); the default is never written to `panew`.
 - **FS-13** **Visibility rides the URL; width does not.** The two halves of pane state are persisted differently on purpose, because they answer different questions.
   - **Visibility → `?preview=true` on the shell URL** (that literal spelling), *plus* the per-folder view state. Toggling writes the param (`replaceSearch`, D8 mechanics); a pane restored from a folder's view state **reflects itself into the URL** so the two never disagree on screen; and on mount the **URL wins** — `?preview` present is authoritative, view state is consulted only in its absence. So the pane is shareable and bookmarkable (SB-2 captures the URL verbatim, and now captures the pane with it) and survives a refresh or a Back into that entry.
   - **Width → per-folder view state only** (`lib/viewstate.ts`, alongside the folder's sort `?sort`/`&order`), never the URL: a bookmark should reproduce *that there is a pane*, not one machine's drag on one window. Two sibling folders keep independent splits. `panew` holds the **fraction** (FS-12) at three decimals — `"0.42"` — and only a **dragged** one is stored; an untouched pane persists nothing. **Legacy pixel values are ignored, not translated:** anything `> 1` in `panew` was written by the pre-fraction model against a container this window may not have, so it reads as absent and the folder opens at the default until it is dragged again. **`pane` and `panew` are independent keys:** turning the pane off clears `pane` but *keeps* `panew`, so closing the pane and then leaving the folder (or refreshing) does not throw away a split the user dragged.
